+++ src/main/java/com/example/examplemod/creativemodetab/ModCreativeModeTabs.java
@@ -44,6 +44,7 @@ public class ModCreativeModeTabs {
                         output.accept(ModItems.AZURITE_BOOTS);
 
                         output.accept(ModItems.AZURITE_HORSE_ARMOR);
+                        output.accept(ModItems.DATA_TABLET);
 
                     }).build());