+++ src/main/java/com/example/examplemod/data/ModDataComponents.java
@@ -1,0 +1,29 @@
+package com.example.examplemod.data;
+
+import java.util.function.UnaryOperator;
+
+import com.example.examplemod.ExampleMod;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.component.DataComponentType;
+import net.minecraft.core.registries.Registries;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.neoforge.registries.DeferredHolder;
+import net.neoforged.neoforge.registries.DeferredRegister;
+
+public class ModDataComponents {
+    public static final DeferredRegister<DataComponentType<?>> DATA_COMPONENT_TYPES = DeferredRegister
+            .createDataComponents(Registries.DATA_COMPONENT_TYPE, ExampleMod.MOD_ID);
+
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<BlockPos>> CORDINATES = register(
+            "cordinates", builder -> builder.persistent(BlockPos.CODEC).networkSynchronized(BlockPos.STREAM_CODEC));
+
+    public static <T> DeferredHolder<DataComponentType<?>, DataComponentType<T>> register(String name,
+            UnaryOperator<DataComponentType.Builder<T>> buildOperator) {
+        return DATA_COMPONENT_TYPES.register(name, () -> buildOperator.apply(DataComponentType.builder()).build());
+    }
+
+    public static void register(IEventBus eventBus) {
+        DATA_COMPONENT_TYPES.register(eventBus);
+    }
+}